requirements.txt
qdrant-client>=1.9.0
transformers>=4.43.0
torch>=2.3.0
pillow>=10.0.0
tqdm>=4.66.0
